+++ src/json_schema/infer.clj
@@ -1,5 +1,42 @@
 (ns json-schema.infer
-  (:require [cheshire.core :as json]))
+  (:require [clojure.walk :as walk]
+            [cheshire.core :as json]))
+
+
+(defn eradicate-lists [m]
+  (if (or (associative? m)
+          (sequential? m))
+    (walk/postwalk
+     #(if (sequential? %)
+        (vec %) %)
+     m)
+    m))
+
+(defn key-paths
+  "Returns a sequence of all key paths in a given map using DFS walk.
+   Converts lists to vectors to enable associative lookup.
+
+   Based on code from:
+   https://dnaeon.github.io/clojure-map-ks-paths/"
+  [m]
+  {:pre [(map? m)]}
+  (let [m (eradicate-lists m)]
+    (letfn [(children [node]
+              (let [v (get-in m node)]
+                (cond (map? v)
+                      (map (fn [x] (conj node x)) (keys v))
+                      (and (sequential? v)
+                           (every? map? v))
+                      (let [d (apply merge (filter map? v))]
+                        (map (fn [x] (conj node 0 x)) (keys d)))
+                      :else [])))
+            (branch? [node] (-> (children node) seq boolean))]
+      (->> (keys m)
+           (map vector)
+           (mapcat (partial tree-seq branch? children))
+           (remove #(and (-> % last keyword?)
+                         (-> % count (not= 1))))
+           vec))))
 
 (defn- init-schema [{:keys [title uri description]}]
   (merge {:$schema "http://json-schema.org/draft-07/schema#"
@@ -13,6 +50,25 @@
   (if (or (keyword? k) (symbol? k))
     (name k)
     (str k)))
+
+(defn- required-path [path]
+  (-> path butlast butlast vec (conj :required)))
+
+(defn- update-required [data {:keys [optional manipulator]} path]
+  (if (-> path count (> 1))
+    (let [k (-> path last)
+          optional? (-> optional (get (keyword k)) boolean)]
+      (update-in data
+                 (required-path path)
+                 (fnil (if optional? disj manipulator) #{})
+                 (sanitize-key k)))
+    data))
+
+(defn- assoc-in-from-either [data path d1 d2]
+  (assoc-in data path
+            (get-in d1 path (get-in d2 path))))
+
+(declare infer) ;; bc/o mutual recursion
 
 (defn infer-strict
   "Naive but strict schema inference from associative data.
@@ -32,7 +88,7 @@
           (or (associative? data)
               (or (nil? title) (string? title)))
           (or (nil? optional) (set? optional)))]}
-  (let [sch (if (and title (not schema))
+  (let [sch (if title
               (init-schema params)
               (or schema {}))]
     (cond (map? data) (merge sch
@@ -45,16 +101,57 @@
                                                                     (dissoc params :title :description :uri))
                                                            second)
                                                      data))
-                                :required (vec (if optional
+                                :required (set (if optional
                                                  (map sanitize-key (remove optional (keys data)))
                                                  sks))}))
           (vector? data) (merge sch
                                 {:type :array
                                  :items (trampoline
-                                         (partial infer-strict
+                                         apply
+                                         (partial infer
                                                   (dissoc params :title :description :uri))
-                                         (first data))})
+                                         data)})
           :else (merge sch (data-type (partial infer-strict (dissoc params :title :description :uri)) data)))))
+
+(defn infer
+  "Schema inference from multiple documents of associative data.
+   Takes n sample documents as input, and returns a schema that allows only
+   data represented in at least one of the documents.
+
+   Data that is present in all documents will be required.
+   Data only represented in a subset of documents will be optional.
+   Data not represented at all, will not be allowed.
+
+   If a single document is passed, infer-strict is used directly.
+
+   Params:
+     optional    - keys that shouldn't be required
+  
+   Optional params:
+     title       - schema title
+     description - schema description
+     uri         - schema uri
+     schema      - continue building on schema"
+  [params & docs]
+  (if (= 1 (count docs))
+    (infer-strict params (first docs))
+    (let [schemas (map (partial infer-strict params) docs)]
+      (reduce
+       (fn [a i]
+         (let [f (frequencies (into (key-paths a) (key-paths i)))]
+           (reduce-kv
+            (fn [a2 k v]
+              (if (= 2 v) ;; required
+                (-> a2
+                    (assoc-in-from-either k a i)
+                    (update-required (assoc params :manipulator conj) k))
+                (-> a2 ;; optional
+                    (assoc-in-from-either k a i)
+                    (update-required (assoc params :manipulator disj) k))))
+            a
+            f)))
+       {}
+       schemas))))
 
 (defn- data-type
   "Return af JSON Schema type map based on input"
@@ -74,9 +171,4 @@
   "A helper function that returns inferred schema as JSON"
   [params data]
   (-> (infer-strict params data) json/encode))
-
-(defn ^:deprecated infer-strict->json
-  "Use infer->schema instead"
-  [data params]
-  (infer->json params data))